+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Copy as HTML on Generated Blog Posts.
 - PDF Generation for blog posts.
+- Centralized location of all AI models used in the app.
 
 
 ## [0.0.7] - 2025-11-01